LayoutTests/media/video-play-pause-exception-expected.txt

Video has no src. Test that play() and pause() rethrow the exception from load(). Test that the play and pause events are not dispatched.

TEST(video.play()) THROWS(DOMException.INVALID_STATE_ERR) OK
TEST(video.pause()) THROWS(DOMException.INVALID_STATE_ERR) OK
END OF TEST
